left encoder broke. :(

+++ src/org/rivierarobotics/subsystems/DriveTrainSide.java
@@ -129,7 +129,7 @@ public class DriveTrainSide {
     }
     
     public double getSidePositionInches() {
-        return masterTalon.getSelectedSensorPosition(0)/ENCODER_CODES_PER_REV*DIST_PER_REV/4.0;
+        return (double)(masterTalon.getSelectedSensorPosition(0))/(double)(ENCODER_CODES_PER_REV)*DIST_PER_REV/4.0;
     }
 
     public double getRawPos() {
